requirements.txt
django==3.2.14
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2